Fix the baseUrl replacement for getSiteUriFromUrl
This ensures that the baseUrl does just get replace once
in the URL for the request. If the path contains the
baseUrl itself it could change the path of the request
and the result would be an unexpected behavior.

In the case of an existing site which is already cached
blitz would return a valid site on a url which should return a 404.
We must ensure that we really only replace the baseUrl and do not
interfere with the path part of the URL.